+++ modules/platforms/cpp/thin-client/include/ignite/thin/transactions/transaction.h
@@ -50,7 +50,10 @@ namespace ignite
                 /**
                  * Destructor.
                  */
-                ~ClientTransaction() {}
+                ~ClientTransaction()
+                {
+                    // No-op.
+                }
 
                 /**
                  * Commits this transaction.
